# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CoopDigital — API de gestion des coopératives cacaoyères
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /livraisons:
    get:
      operationId: getLivraisons
      tags: [livraisons]
      summary: Liste des livraisons
      parameters:
        - name: membre_id
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Liste des livraisons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LivraisonDetail"
    post:
      operationId: createLivraison
      tags: [livraisons]
      summary: Enregistrer une livraison avec déduction d'avance automatique
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LivraisonInput"
      responses:
        "201":
          description: Livraison enregistrée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LivraisonResult"
        "400":
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/parts-sociales/config:
    get:
      operationId: getConfigParts
      tags: [parts-sociales]
      summary: Config parts sociales de la coopérative
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigPartsSociales"
    put:
      operationId: updateConfigParts
      tags: [parts-sociales]
      summary: Modifier la config parts sociales
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigPartsSociales"
      responses:
        "200":
          description: Config mise à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigPartsSociales"
  /api/parts-sociales/rapport:
    get:
      operationId: getRapportParts
      tags: [parts-sociales]
      summary: Rapport global parts sociales
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Rapport
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RapportParts"
  /api/parts-sociales/membre/{id}:
    get:
      operationId: getPartsMembre
      tags: [parts-sociales]
      summary: Parts sociales d'un membre
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Parts du membre
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartsMembre"
  /api/parts-sociales/liberation:
    post:
      operationId: enregistrerLiberation
      tags: [parts-sociales]
      summary: Enregistrer un versement de libération
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LiberationInput"
      responses:
        "201":
          description: Versement enregistré
          content:
            application/json:
              schema:
                type: object
                required: [succes]
                properties:
                  succes:
                    type: boolean
...
  /api/paiements:
    get:
      operationId: listPaiements
      tags: [paiements]
      summary: Liste des paiements producteurs
      security: [{ bearerAuth: [] }]
      parameters:
        - name: statut
          in: query
          schema:
            type: string
            enum: [en_attente, confirme, echec]
        - name: membre_id
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Liste paiements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaiementListItem"
  /api/paiements/{id}/valider:
    patch:
      operationId: validerPaiement
      tags: [paiements]
      summary: Marquer un paiement comme confirmé
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValiderPaiementInput"
      responses:
        "200":
          description: Paiement confirmé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaiementListItem"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [erreur]
      properties:
        erreur:
          type: string
...
    Avance:
      type: object
      required: [id, membreId, montantOctroyeFcfa, montantRembourseFcfa, soldeRestantFcfa, dateOctroi, statut, createdAt]
      properties:
        id:
          type: integer
        membreId:
          type: integer
        montantOctroyeFcfa:
          type: integer
        montantRembourseFcfa:
          type: integer
        soldeRestantFcfa:
          type: integer
        dateOctroi:
          type: string
        dateEcheance:
          type: ["string", "null"]
        motif:
          type: ["string", "null"]
        statut:
          type: string
          enum: [en_cours, rembourse, en_retard]
        agentId:
          type: ["integer", "null"]
        createdAt:
          type: string
        membreNom:
          type: ["string", "null"]
        membrePrenoms:
          type: ["string", "null"]
...
    LivraisonInput:
      type: object
      required: [membreId, poidsKg, prixUnitaireFcfa, dateLivraison]
      properties:
        membreId:
          type: integer
        poidsKg:
          type: number
        prixUnitaireFcfa:
          type: integer
        dateLivraison:
          type: string
        modePaiement:
          type: string
          enum: [orange_money, mtn_momo, especes]
        campagneId:
          type: integer
          nullable: true
        nombreSacs:
          type: integer
          nullable: true
        retenueKg:
          type: number
          nullable: true
        sectionLivraison:
          type: string
          nullable: true
    ValiderPaiementInput:
      type: object
      properties:
        referenceTransaction:
          type: string
          nullable: true
        dateReglement:
          type: string
          nullable: true
    PaiementListItem:
      type: object
      required: [id, livraisonId, membreId, montantFcfa, modePaiement, statut, createdAt]
      properties:
        id:
          type: integer
        livraisonId:
          type: integer
        membreId:
          type: integer
        montantFcfa:
          type: integer
        modePaiement:
          type: string
          enum: [orange_money, mtn_momo, especes]
        referenceTransaction:
          type: ["string", "null"]
        statut:
          type: string
          enum: [en_attente, confirme, echec]
        createdAt:
          type: string
        membreNom:
          type: ["string", "null"]
        membrePrenoms:
          type: ["string", "null"]
        telephone:
          type: ["string", "null"]
        dateLivraison:
          type: ["string", "null"]
    LivraisonDetail:
      type: object
      required: [id, membreId, poidsKg, prixUnitaireFcfa, montantBrutFcfa, avanceDeduiteFcfa, montantNetFcfa, dateLivraison, createdAt]
      properties:
        id:
          type: integer
        membreId:
          type: integer
        poidsKg:
          type: string
        prixUnitaireFcfa:
          type: integer
        montantBrutFcfa:
          type: integer
        avanceDeduiteFcfa:
          type: integer
        montantNetFcfa:
          type: integer
        dateLivraison:
          type: string
        agentId:
          type: ["integer", "null"]
        createdAt:
          type: string
        membreNom:
          type: ["string", "null"]
        membrePrenoms:
          type: ["string", "null"]
    LivraisonResult:
      type: object
      required: [livraison, paiement]
      properties:
        livraison:
          $ref: "#/components/schemas/LivraisonDetail"
        paiement:
          $ref: "#/components/schemas/Paiement"
        avanceMiseAJour:
          $ref: "#/components/schemas/Avance"
    Paiement:
      type: object
      required: [id, livraisonId, membreId, montantFcfa, modePaiement, statut, recuEnvoyeWhatsapp, createdAt]
      properties:
        id:
          type: integer
        livraisonId:
          type: integer
        membreId:
          type: integer
        montantFcfa:
          type: integer
        modePaiement:
          type: string
          enum: [orange_money, mtn_momo, especes]
        referenceTransaction:
          type: ["string", "null"]
        statut:
          type: string
          enum: [en_attente, confirme, echec]
        recuEnvoyeWhatsapp:
          type: boolean
        createdAt:
          type: string
...
    ConfigPartsSociales:
      type: object
      required: [cooperativeId, valeurNominaleFcfa, nbrePartsMin]
      properties:
        cooperativeId:
          type: integer
        valeurNominaleFcfa:
          type: integer
        nbrePartsMin:
          type: integer
    RapportParts:
      type: object
      properties:
        totalMembresSouscripteurs:
          type: ["integer", "null"]
        totalPartsSouscrites:
          type: ["integer", "null"]
        totalSouscritFcfa:
          type: ["integer", "null"]
        totalLibereFcfa:
          type: ["integer", "null"]
        totalResteALibererFcfa:
          type: ["integer", "null"]
    LiberationParts:
      type: object
      required: [id, membreId, cooperativeId, dateVersement, montantFcfa, createdAt]
      properties:
        id:
          type: integer
        membreId:
          type: integer
        cooperativeId:
          type: integer
        dateVersement:
          type: string
          format: date
        codeLiberation:
          type: ["string", "null"]
        versement:
          type: ["string", "null"]
        montantFcfa:
          type: integer
        agentId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    PartsMembre:
      type: object
      required: [membre, liberations, config]
      properties:
        membre:
          type: object
          required: [id, nom, prenoms, nbrePartsSouscrites, valeurNominalePartFcfa, totalSouscritFcfa, totalLibereFcfa, resteALibererFcfa]
          properties:
            id:
              type: integer
            nom:
              type: string
            prenoms:
              type: string
            nbrePartsSouscrites:
              type: integer
            valeurNominalePartFcfa:
              type: integer
            totalSouscritFcfa:
              type: integer
            totalLibereFcfa:
              type: integer
            resteALibererFcfa:
              type: integer
        liberations:
          type: array
          items:
            $ref: "#/components/schemas/LiberationParts"
        config:
          $ref: "#/components/schemas/ConfigPartsSociales"
    LiberationInput:
      type: object
      required: [membreId, montantFcfa, dateVersement]
      properties:
        membreId:
          type: integer
        montantFcfa:
          type: integer
        dateVersement:
          type: string
          format: date
        codeLiberation:
          type: string
        versement:
          type: string
...