# lib/api-zod/src/generated/types/liberationParts.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CoopDigital — API de gestion des coopératives cacaoyères
 * OpenAPI spec version: 0.1.0
 */

export interface LiberationParts {
  id: number;
  membreId: number;
  cooperativeId: number;
  dateVersement: Date;
  /** @nullable */
  codeLiberation?: string | null;
  /** @nullable */
  versement?: string | null;
  montantFcfa: number;
  /** @nullable */
  agentId?: number | null;
  createdAt: Date;
}