# lib/api-zod/src/generated/types/rapportParts.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CoopDigital — API de gestion des coopératives cacaoyères
 * OpenAPI spec version: 0.1.0
 */

export interface RapportParts {
  /** @nullable */
  totalMembresSouscripteurs?: number | null;
  /** @nullable */
  totalPartsSouscrites?: number | null;
  /** @nullable */
  totalSouscritFcfa?: number | null;
  /** @nullable */
  totalLibereFcfa?: number | null;
  /** @nullable */
  totalResteALibererFcfa?: number | null;
}